django_typeBiet/django_api1/requirements.txt
django==3.0
djangorestframework
markdown==3.0.0
django-filter==1.0.1
python-decouple==3.1
Pillow==6.0.0